+++ src/Core/GUI/SwingEX/XConstants.java
@@ -1,0 +1,20 @@
+package Core.GUI.SwingEX;
+
+/**
+ * KM
+ * JUne 04 2017
+ * Handles the constants for SwingEX
+ */
+
+public interface XConstants {
+
+    int HORIZONTAL_SORT = 1;
+    int VERTICAL_SORT = 2;
+    int ORIENT_LEFT = 3;
+    int ORIENT_RIGHT = 4;
+    int ORIENT_TOP = 5;
+    int ORIENT_BOTTOM = 6;
+    int ORIENT_CENTER = 7;
+
+
+}